Fix field masks for presence with default value #344
